Module hllvm::ir

The LLVM intermediate representation.

Reexports

pub use self::context::Context;

pub use self::module::Module;

pub use self::attribute::Attribute;

pub use self::ty::*;

pub use self::value::*;

pub use sys::ir::opcode::*;

Modules

attribute
An attribute which can be placed on certain IR data structures.

context
The main LLVM context type.

macros
Useful macros.

module
The module-related code.

ty
Definitions for all of the different types.

value

Enums

AtomicBinaryOp
Integer predicate kind.

AtomicOrdering
Atomic ordering.

FloatPredicateKind
Float predicate kind.

IntegerPredicateKind
Integer predicate kind.

Linkage
An enumeration for the kinds of linkage for global values.

SynchronizationScope
Synchronization scope.

ThreadLocalMode
Thread local mode.